LayoutTests/fast/animation/request-animation-frame-timestamps-expected.txt

Tests the timestamps provided to requestAnimationFrame callbacks

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


In firstRequestAnimationFrame()
In secondRequestAnimationFrame()
PASS firstTimestamp is secondTimestamp
PASS successfullyParsed is true

TEST COMPLETE
